Imported files from old Visual Frameworks app.
While this app is ugly, it is pure HTML, CSS and Javascript, and has
CRUD functionality. All storage is Local Storage, but it works. Over
the next three weeks this will be modified to work with jQuery, jQuery
Mobile and possibly CouchDB for storage as well as be modified to work
with PhoneGap to provide access to "native" Android and iOS functions.